honor cancel messages (according to policy), keeping track of what should be cancelled. the gui causes one or two cancel messages to be created when the cancel option is selected - first, it creates a specific stub cancel message containing no data, just a header with the to-be-cancelled uri, and second, it (may) create an updated metadata to include the to-be-cancelled uri in its list of recently cancelled messages.
the stub message is authorized and authenticated as the author of the to-be-cancelled
message, or if the author's keys are not known, it is authorized and authenticated by
the owner of the forum the to-be-cancelled message was posted in.  if neither of those
keys were known, no cancel message is created (down the line, a gui for some other
options can be added).  the metadata to be upated uses that same policy - it updates
the author's metadata (if possible) or the forum's metadata (if possible).